pyproject.toml
[tool.poetry]
name = "snapstream"
version = "0.0.0.dev0"
description = "Streamline your Kafka data processing, this tool aims to standardize streaming data from multiple Kafka clusters. With a pub-sub approach, multiple functions can easily subscribe to incoming messages, serialization can be specified per topic, and data is automatically processed by data sink functions."
authors = ["Menziess <stefan_schenk@hotmail.com>"]
readme = "README.md"
homepage = "https://github.com/Menziess/snapstream"

[tool.poetry.dependencies]
python = "^3.8.1"
confluent-kafka = "^2.0.2"
rocksdict = "^0.3.10"
pypubsub = "^4.0.3"
avro = "^1.11.1"
toolz = "^0.12.0"
pyright = "^1.1.302"

[tool.poetry.group.dev.dependencies]
pydocstyle = "^6.3.0"
autopep8 = "^2.0.2"
pyright = "^1.1.300"
flake8 = "^6.0.0"
bandit = "^1.7.5"
pytest = "^7.2.2"
pytest-cov = "^4.0.0"
pre-commit = "^3.2.1"
pytest-mock = "^3.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pyright]
include = ["snapstream"]
venvPath = "."
venv = ".venv"

[tool.bandit]
exclude_dirs = ["tests"]

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = [
    "snapstream",
    "tests",
]
